refactor: Move all files to package "cheems"
