+++ src/main/java/cheems/Cheems.java
@@ -1,3 +1,5 @@
+package cheems;
+
 import java.util.Scanner;
 
 public class Cheems {
