+++ src/main/java/cheems/Keyword.java
@@ -1,3 +1,5 @@
+package cheems;
+
 public enum Keyword {
     MARK,
     UNMARK,
