+++ src/main/java/cheems/exceptions/IOException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package cheems.exceptions;
 
 public class IOException extends RuntimeException {
     public IOException(String msg) {
