+++ src/main/java/cheems/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package tasks;
+package cheems.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
